improve(sdk): move prices table to be historical prices
