+++ packages/sdk/src/libs/index.ts
@@ -2,6 +2,7 @@ export * as clients from "./clients";
 export * as stores from "./stores";
 export * as tables from "./tables";
 export * as utils from "./utils";
+export { default as Coingecko } from "./coingecko";
 
 // types
 import type { TypedEventFilter, TypedEvent } from "@uma/core/contract-types/ethers/commons";
