+++ packages/sdk/src/libs/tables/generic/sorted-js-map.ts
@@ -1,8 +1,7 @@
 import Base from "../base";
-import Store from "../../stores/sorted-js-map";
+import { SortedStore, SortedJsMap, Store } from "../../stores";
 
 export default function <I, D>(type: string, makeId: (d: D) => I) {
-  const store = Store<I, D>();
-  const table = Base<I, D>({ type, makeId }, store);
-  return table;
+  const store = SortedJsMap<I, D>();
+  return Base<I, D>({ type, makeId }, store);
 }
