+++ packages/sdk/src/libs/tables/historical-prices/index.ts
