+++ packages/sdk/src/libs/tables/historical-prices/sorted-js-map.ts
@@ -1,7 +1,7 @@
-import { Data, makeId } from ".";
+import { Data, makeId } from "./utils";
 import { SortedJsMap } from "../generic";
 
-export const Table = (type = "Price") => {
+export const Table = (type = "Historical Price") => {
   return SortedJsMap<string, Data>(type, makeId);
 };
 export type Table = ReturnType<typeof Table>;
