+++ packages/sdk/src/libs/tables/historical-prices/utils.ts
