+++ packages/sdk/src/libs/tables/index.ts
@@ -2,4 +2,4 @@ export { default as base } from "./base";
 export * as blocks from "./blocks";
 export * as emps from "./emps";
 export * as generic from "./generic";
-export * as prices from "./prices";
+export * as historicalPrices from "./historical-prices";
